refactor: move NewYear threshold logic to a NewYears class
